+++ lib/utils.ts
@@ -39,3 +39,10 @@ ${json.meta.errors.map((error: string) => `  - ${error}`).join('\n')}`;
     this.body = body;
   }
 }
+
+export class MissingParamError extends Error {
+  constructor(param: string) {
+    super(param);
+    this.message = `${param} is required`;
+  }
+}